Fetch scene after FrameSort and Frame Ignore toggle to refresh FramePreviewCard.

+++ frontend/static/frontend/startDemo.js
@@ -240,6 +240,10 @@ if (window.localStorage){
 
 
 // for debug
-document.querySelector("#testreset").onclick=(e)=>{
-    window.localStorage.clear();
+const $testReset = document.querySelector("#testreset");
+if ($testReset){
+    $testReset.onclick=(e)=>{
+        window.localStorage.clear();
+    }
 }
+
